requirements.txt
fastapi
uvicorn
python-multipart
pillow
ultralytics
opencv-python
numpy
requests
pyngrok
tqdm
seaborn

# pip install -r requirements.txt รันทุกครั้งเมื่อย้ายเครื่อง
